+++ Gw2Sharp/IGw2Client.cs
@@ -1,3 +1,5 @@
+using System;
+using Gw2Sharp.Mumble;
 using Gw2Sharp.WebApi;
 
 namespace Gw2Sharp
@@ -5,8 +7,13 @@ namespace Gw2Sharp
     /// <summary>
     /// An interface for the client implementation of Guild Wars 2 APIs.
     /// </summary>
-    public interface IGw2Client
+    public interface IGw2Client : IDisposable
     {
+        /// <summary>
+        /// Gets the Mumble Link API.
+        /// </summary>
+        IGw2MumbleClient Mumble { get; }
+
         /// <summary>
         /// Gets the web API.
         /// </summary>